feat: 사용자 운전 점수 조회 API 개발

+++ drivecare/src/main/java/com/gitauto/drivecare/database/user_driving_stat/entity/UserDrivingStatEntity.java
@@ -25,7 +25,7 @@ public class UserDrivingStatEntity {
     private UserInfoEntity userInfo;
 
     @Column(name = "DRIVE_SCORE")
-    private Long driveScore;
+    private Double driveScore;
 
     @Column(name = "ACCEL_COUNT")
     private Integer accelCount;
